logaddexp2: Use log1p and exp2
This replaces `log2(1 + x)` with `log1p(x) * (1 / log(2))` which improves
precision when `x` is small by avoiding the truncation from calculating
`(1 + x) - 1`. Noting that `x` is always `<= 1` in this formula.

This also replaces `pow(2, x)` with `exp2(x)` which improves performance,
particularly on CPU where the constant value cannot be inlined into Sleef.
With numel=1e7 for example, I see a 1.35x speedup on CPU.

ghstack-source-id: 2d3a2a7e5b45c41c2f4116a3a48bb4d52c6e6ae6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/92116